requirements.txt
opencv-python
pillow
requests
python-dotenv